# docs/developer/api.yaml
...
openapi: 3.1.0
info:
  title: OpenTalk Controller Frontend API
  description: Specifies the endpoints and structure of the OpenTalk Controller Frontend API
  contact:
    name: OpenTalk Team
    email: mail@opentalk.eu
  license:
    name: EUPL-1.2
    identifier: EUPL-1.2
  version: 0.35.0-alpha
servers:
  - url: /v1
paths:
...
  "/internal/room/{room_id}/asset":
    post:
      tags:
        - "api::internal::assets"
      summary: |-
        Upload an asset. This endpoint requires the client to provide the RoomServer
        credentials.
      operationId: internal_upload_asset
      parameters:
        - name: room_id
          in: path
          description: The id of the room
          required: true
          schema:
            $ref: "#/components/schemas/RoomId"
        - name: event_title
          in: query
          description: The title of the event
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/EventTitle"
        - name: file_extension
          in: query
          description: The file extension
          required: true
          schema:
            $ref: "#/components/schemas/FileExtension"
        - name: kind
          in: query
          description: The kind of the asset
          required: true
          schema:
            $ref: "#/components/schemas/AssetFileKind"
        - name: namespace
          in: query
          description: The namespace the asset is assigned to
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/ModuleId"
      requestBody:
        description: The contents of the file
        content:
          application/octet-stream:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: The asset has been created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAssetResponseBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
...
  "/rooms/{room_id_or_alias}/assets":
    get:
      tags:
        - "api::v1::assets"
      summary: Get the assets associated with a room.
      description: |-
        This returns assets that are available for a room. If no
        pagination query is added, the default page size is used.
      operationId: room_assets
      parameters:
        - name: room_id_or_alias
          in: path
          description: Either the id or the alias of the room
          required: true
          schema:
            $ref: "#/components/schemas/RoomIdOrAlias"
        - name: per_page
          in: query
          description: The number of entries per page
          required: false
          schema:
            $ref: "#/components/schemas/PageSize"
        - name: page
          in: query
          description: The number of the page
          required: false
          schema:
            $ref: "#/components/schemas/Page"
      responses:
        "200":
          description: The assets have been returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomsByRoomIdAssetsGetResponseBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
    post:
      tags:
        - "api::v1::assets"
      summary: Create an asset for a room from an uploaded file
      description: The asset is attached to the room and saved in the storage.
      operationId: create_room_asset
      parameters:
        - name: room_id_or_alias
          in: path
          description: Either the id or the alias of the room
          required: true
          schema:
            $ref: "#/components/schemas/RoomIdOrAlias"
        - name: event_title
          in: query
          description: The title of the event
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/EventTitle"
        - name: file_extension
          in: query
          description: The file extension
          required: true
          schema:
            $ref: "#/components/schemas/FileExtension"
        - name: kind
          in: query
          description: The kind of the asset
          required: true
          schema:
            $ref: "#/components/schemas/AssetFileKind"
        - name: namespace
          in: query
          description: The namespace the asset is assigned to
          required: false
          schema:
            oneOf:
              - type: "null"
              - $ref: "#/components/schemas/ModuleId"
      requestBody:
        description: The contents of the file
        content:
          application/octet-stream:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: The asset has been created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAssetResponseBody"
        "400":
          description: Storage quota has been exceeded
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: The associated room was not found
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    AssetFileKind:
      type: string
      description: An asset file kind
      examples:
        - mykind
      maxLength: 50
      minLength: 1
      pattern: "^[0-9a-zA-Z_]*$"
    AssetId:
      type: string
      format: uuid
      description: The id of an asset
      example: 00000000-0000-0000-0000-0000aabbcc00
    AssetResource:
      type: object
      description: Representation of an asset resource
      required:
        - id
        - filename
        - created_at
        - kind
        - size
      properties:
        created_at:
          type: string
          format: date-time
          description: The timestamp the asset was created
        filename:
          type: string
          description: The filename of the asset
        id:
          $ref: "#/components/schemas/AssetId"
          description: The ID of an asset
        kind:
          type: string
          description: The asset kind
        namespace:
          $ref: "#/components/schemas/ModuleId"
          description: The namespace of the asset
        size:
          $ref: "#/components/schemas/FileSize"
          description: The size of the asset in bytes
      example:
        created_at: "2024-06-18T11:22:33Z"
        filename: recording.webm
        id: 00000000-0000-0000-0000-0000aabbcc00
        kind: record
        namespace: recording
        size: 98765432
...
    EventTitle:
      type: string
      description: The title of an event
      examples:
        - Team Event
      maxLength: 255
...
    FileExtension:
      type: string
      description: An extension for a file path
      examples:
        - pdf
      maxLength: 10
      pattern: "^[0-9a-zA-Z]*$"
    FileSize:
      type: integer
      format: int64
      description: The size of a file.
      example: 17
...
    ModuleId:
      anyOf:
        - type: string
          description: A module identifier
          examples:
            - mymodule
          maxLength: 255
          minLength: 1
          pattern: "^[_0-9a-z]*$"
      default: core
      example: mymodule
...
    Page:
      type: integer
      format: int64
      description: |-
        A page number used for paging in requests.

        This is `1`-based, so a `Page` with a value of `0` can not be created.
      example: 5
...
    PageSize:
      type: integer
      format: int64
      description: A page number used for paging in requests.
      example: 30
...
    PostAssetResponseBody:
      type: object
      description: "Response for the `POST /services/rooms/{room_id}/asset` endpoint"
      required:
        - asset_resource
        - quota
      properties:
        asset_resource:
          $ref: "#/components/schemas/AssetResource"
          description: The asset that was uploaded
        quota:
          $ref: "#/components/schemas/Quota"
          description: The total and used quota in bytes
...
    Quota:
      type: object
      description: Quota information
      required:
        - used
      properties:
        total:
          type:
            - integer
            - "null"
          format: int64
          description: "The total quota available, if applicable"
          minimum: 0
        used:
          type: integer
          format: int64
          description: The used quota
          minimum: 0
...
    RoomAlias:
      type: string
      description: A human readable alias for a room.
      examples:
        - personal-room-name_0000000000000000
    RoomId:
      type: string
      format: uuid
      description: The id of a room
      example: 00000000-0000-0000-0000-0000abadcafe
    RoomIdOrAlias:
      oneOf:
        - $ref: "#/components/schemas/RoomId"
          description: A room id.
        - $ref: "#/components/schemas/RoomAlias"
          description: A room alias.
      description: An enum that represents either a room id or a room alias.
      example: 00000000-0000-0000-0000-0000abadcafe
...
    RoomsByRoomIdAssetsGetResponseBody:
      type: array
      items:
        $ref: "#/components/schemas/AssetResource"
      description: "Response for *GET /rooms/{room_id}/assets*"
      example:
        - created_at: "2024-06-18T11:22:33Z"
          filename: recording.webm
          id: 00000000-0000-0000-0000-0000aabbcc00
          kind: record
          namespace: recording
          size: 98765432
...
  responses:
...
    Forbidden:
      description: The authorized user has no permission to access the requested resource
    InternalServerError:
      description: An internal server error occurred
    NotFound:
      description: The requested data could not be found
    Unauthorized:
      description: |-
        The provided access token is expired or the provided id or access token is invalid.
                The WWW-Authenticate header will contain an error description 'session expired' to distinguish between
                an invalid and an expired token
      headers:
        www-authenticate:
          schema:
            type: string
          description: |

            Will contain 'session expired' to distinguish between an invalid and an expired token.

            Examples:

                Bearer error="invalid_token", error_description="The provided access token is invalid"
                Bearer error="invalid_token", error_description="The user session expired"
      content:
        application/json:
          schema:
            type: object
            description: Internal reusable dummy type for utoipa unauthorized error
            required:
              - code
              - message
            properties:
              code:
                type: string
                description: Machine readable error code
              message:
                type: string
                description: Human readable message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    InviteCode:
      type: http
      scheme: bearer
...